chore: move copyObject, httpGet, httpPost
