+++ common/utils.js
@@ -2,6 +2,8 @@
 
 const { outputJsonSync } = require('fs-extra');
 const chalk = require('chalk');
+const https = require('https');
+const http = require('http');
 
 function loadConfig(env) {
     return require(`${__dirname}/../axelar-chains-config/info/${env}.json`);
@@ -96,6 +98,57 @@ const isNonEmptyStringArray = (arr) => {
     return true;
 };
 
+function copyObject(obj) {
+  return JSON.parse(JSON.stringify(obj));
+}
+
+const httpGet = (url) => {
+  return new Promise((resolve, reject) => {
+      (url.startsWith('https://') ? https : http).get(url, (res) => {
+          const { statusCode } = res;
+          const contentType = res.headers['content-type'];
+          let error;
+
+          if (statusCode !== 200 && statusCode !== 301) {
+              error = new Error('Request Failed.\n' + `Request: ${url}\nStatus Code: ${statusCode}`);
+          } else if (!/^application\/json/.test(contentType)) {
+              error = new Error('Invalid content-type.\n' + `Expected application/json but received ${contentType}`);
+          }
+
+          if (error) {
+              res.resume();
+              reject(error);
+              return;
+          }
+
+          res.setEncoding('utf8');
+          let rawData = '';
+          res.on('data', (chunk) => {
+              rawData += chunk;
+          });
+          res.on('end', () => {
+              try {
+                  const parsedData = JSON.parse(rawData);
+                  resolve(parsedData);
+              } catch (e) {
+                  reject(e);
+              }
+          });
+      });
+  });
+};
+
+const httpPost = async (url, data) => {
+  const response = await fetch(url, {
+      method: 'POST',
+      headers: {
+          'Content-Type': 'application/json',
+      },
+      body: JSON.stringify(data),
+  });
+  return response.json();
+}
+
 module.exports = {
     loadConfig,
     saveConfig,
@@ -112,4 +165,7 @@ module.exports = {
     isValidDecimal,
     isNumberArray,
     isNonEmptyStringArray,
+    copyObject,
+    httpGet,
+    httpPost,
 };
